absorb EndOnConditionCommand into WaitCommand
