+++ wpilibj/src/test/java/edu/wpi/first/wpilibj/experimental/command/CommandRequirementsTest.java
@@ -269,7 +269,7 @@ public class CommandRequirementsTest extends CommandTestBase {
 
     Subsystem system = new TestSubsystem();
 
-    Command missingRequirement = new EndOnConditionCommand(() -> false);
+    Command missingRequirement = new WaitUntilCommand(() -> false);
     Command ends = new InstantCommand(() -> { }, system);
 
     assertThrows(IllegalUseOfCommandException.class,
